refactor: SuccessTextBox into folder
refactor: update success text box and variables css
- uses style to remove need for chakra in component
