+++ src/styles/variables.css
@@ -17,6 +17,7 @@
 
     /* semantic colors  */
     --amp-shadows-sm: 0 1px 2px 0 rgba(0, 0, 0, 0.05);
+    --amp-shadows-md: 0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06);
     --amp-colors-border: var(--amp-colors-neutral-200);
     --amp-colors-error-background: #FEF2F2;
     --amp-colors-error-border: #FECACA;
